# Optimize branch creation