@aome510 2022-07-14
The speedup can be measured by running the `test_branch_creation_heavy_write` test against the PR and `main`. At least based on runs in my local laptop, the maximum latency for this PR should be 3-5x smaller than what of `main`.