@aome510 2022-07-15
> This doesn't look safe to me. This can happen:
> 
> In the beginning, there is only one branch, 'main'. Thread A calls `branch_timeline('main', 'new', '1/11112222)` to create new branch called 'new', at LSN '1/10000000`. At the same time, thread B calls gc_iteration_internal().
> 
> 1. A acquires layer_removal_cs
> 2. A reads latest_gc_cutoff_lsn as 1/11110000.
> 3. B scans the list of timelines. The new timeline creation hasn't finished yet, so it only sees 'main'
> 4. B calls gc() on the source timeline. It blocks on layer_removal_cs
> 5. A finishes creating the new timeline
> 6. B goes ahead with the garbage collection, updates latest_gc_cutoff_lsn to 1/22220000, and removes layers older than that.
> 
> Step 6 will remove data that is still needed by the new timeline.
> 
> In a nutshell, GC needs to be careful to not remove data that is still needed by child branches. There's a race condition between new branch creation and GC. If a new branch is created after GC has collected its list of branches and their branch-points (`retain_lsns`), but before it has updated `latest_gc_cutoff_lsn` on the source timeline, it can remove data that is still needed by the new timeline.

Thanks for taking a look into this. I have separated the GC codes into 2 phases as you described in the later comment. Also added a test `test_branch_creation_before_gc` simulating the above scenario.

Still need to do some cleanups such as adding more comments and updating the documents