@aome510 2022-07-19
Result when running the new branch creation perf tests comparing with the latest `main`:

This patch:
```
test_branch_creation_heavy_write[20].branch_creation_duration_max: 3.961 s
test_branch_creation_heavy_write[20].branch_creation_duration_avg: 1.185 s
test_branch_creation_heavy_write[20].branch_creation_duration_stdev: 1.222 s
test_branch_creation_many[1024].branch_creation_duration_max: 0.273 s
test_branch_creation_many[1024].branch_creation_duration_avg: 0.102 s
test_branch_creation_many[1024].branch_creation_duration_stdev: 0.037 s
```

main:
```
test_branch_creation_heavy_write[20].branch_creation_duration_max: 21.256 s
test_branch_creation_heavy_write[20].branch_creation_duration_avg: 2.072 s
test_branch_creation_heavy_write[20].branch_creation_duration_stdev: 5.348 s
test_branch_creation_many[1024].branch_creation_duration_max: 3.733 s
test_branch_creation_many[1024].branch_creation_duration_avg: 0.106 s
test_branch_creation_many[1024].branch_creation_duration_stdev: 0.119 s
```

In general, looks like a decent improvement. Will merge once tests passed